requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
faker
matplotlib